+++ deps.ts
@@ -3,12 +3,3 @@ export * as ConsoleColor from "https://deno.land/x/colorlog@v1.0/mod.ts";
 export { default as SQLQueryBuilder } from "https://raw.githubusercontent.com/denjucks/dex/master/mod.ts";
 
 export { camelCase, snakeCase } from "https://deno.land/x/case/mod.ts";
-
-export { Client as PostgresClient } from "https://raw.githubusercontent.com/deno-postgres/deno-postgres/master/mod.ts";
-
-export {
-  Client as MySQLClient,
-  Connection as MySQLConnection,
-} from "https://deno.land/x/mysql/mod.ts";
-
-export { DB as SQLiteClient } from "https://deno.land/x/sqlite@v2.3.1/mod.ts";
