+++ lib/connectors/mysql/connector.ts
@@ -1,9 +1,9 @@
-import { MySQLClient, MySQLConnection } from "../../deps.ts";
-import type { Connector, ConnectorOptions } from "./connector.ts";
-import { SQLTranslator } from "../translators/sql-translator.ts";
-import type { QueryDescription } from "../query-builder.ts";
-import { DatabaseDialect } from "../database.ts";
-import { Translator } from "../translators/translator.ts";
+import { MySQLClient, MySQLConnection } from "./deps.ts";
+import type { Connector, ConnectorOptions } from "../connector.ts";
+import { SQLTranslator } from "../../translators/sql-translator.ts";
+import type { QueryDescription } from "../../query-builder.ts";
+import { BuiltInDatabaseDialect } from "../../database.ts";
+import { Translator } from "../../translators/translator.ts";
 
 export interface MySQLOptions extends ConnectorOptions {
   database: string;
@@ -14,7 +14,7 @@ export interface MySQLOptions extends ConnectorOptions {
 }
 
 export class MySQLConnector implements Connector {
-  _dialect: DatabaseDialect = 'mysql';
+  _dialect: BuiltInDatabaseDialect = "mysql";
 
   _client: MySQLClient;
   _options: MySQLOptions;
