+++ lib/connectors/mysql/deps.ts
@@ -1,0 +1,4 @@
+export {
+  Client as MySQLClient,
+  Connection as MySQLConnection,
+} from "https://deno.land/x/mysql/mod.ts";
