+++ lib/connectors/postgres/connector.ts
@@ -1,9 +1,9 @@
-import { PostgresClient } from "../../deps.ts";
-import type { Connector, ConnectorOptions } from "./connector.ts";
-import { SQLTranslator } from "../translators/sql-translator.ts";
-import type { QueryDescription } from "../query-builder.ts";
-import type { Values } from "../data-types.ts";
-import { DatabaseDialect } from "../database.ts";
+import { PostgresClient } from "./deps.ts";
+import type { Connector, ConnectorOptions } from "../connector.ts";
+import { SQLTranslator } from "../../translators/sql-translator.ts";
+import type { QueryDescription } from "../../query-builder.ts";
+import type { Values } from "../../data-types.ts";
+import { BuiltInDatabaseDialect } from "../../database.ts";
 
 export interface PostgresOptions extends ConnectorOptions {
   database: string;
@@ -14,7 +14,7 @@ export interface PostgresOptions extends ConnectorOptions {
 }
 
 export class PostgresConnector implements Connector {
-  _dialect: DatabaseDialect = 'postgres';
+  _dialect: BuiltInDatabaseDialect = "postgres";
 
   _client: PostgresClient;
   _options: PostgresOptions;
