+++ mod.ts
@@ -1,4 +1,5 @@
 export { DATA_TYPES, DataTypes } from "./lib/data-types.ts";
-export { Database } from "./lib/database.ts";
+export { Database, DatabaseOptions, SyncOptions } from "./lib/database.ts";
 export { Model } from "./lib/model.ts";
 export { Relationships } from "./lib/relationships.ts";
+export * from "./lib/connectors/";
